feat: ReservationService 인터페이스 추가
